- Removed hashtable. - Replaced uint and int with size_t in some places, where appropriate. - Hashed string now usees uint64

+++ Source/Engine/Foundation/Types/VanguardList.h
@@ -25,7 +25,7 @@ namespace Vanguard
 			data.push_back(aItem);
 		}
 
-		inline uint32 Size() const
+		inline size_t Size() const
 		{
 			return data.size();
 		}
